# cmd/apigen/gen/http/inventory/server/encode_decode.go
// Code generated by goa, DO NOT EDIT.
//
// inventory HTTP server encoders and decoders
//
// Command:
// $ goa gen github.com/eclipse-xfsc/kubernetes-operator/design

package server

import (
	"context"
	"net/http"

	inventory "github.com/eclipse-xfsc/kubernetes-operator/cmd/apigen/gen/inventory"
	inventoryviews "github.com/eclipse-xfsc/kubernetes-operator/cmd/apigen/gen/inventory/views"
	goahttp "goa.design/goa/v3/http"
)

// EncodeVersionResponse returns an encoder for responses returned by the
// inventory version endpoint.
func EncodeVersionResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		res := v.(*inventoryviews.XsfcVersion)
		enc := encoder(ctx, w)
		body := NewVersionResponseBody(res.Projected)
		w.WriteHeader(http.StatusOK)
		return enc.Encode(body)
	}
}

// EncodeModulesResponse returns an encoder for responses returned by the
// inventory modules endpoint.
func EncodeModulesResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		res, _ := v.([]*inventory.XsfcModule)
		enc := encoder(ctx, w)
		body := NewModulesResponseBody(res)
		w.WriteHeader(http.StatusOK)
		return enc.Encode(body)
	}
}

// EncodeTypesResponse returns an encoder for responses returned by the
// inventory types endpoint.
func EncodeTypesResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		res, _ := v.([]string)
		enc := encoder(ctx, w)
		body := res
		w.WriteHeader(http.StatusOK)
		return enc.Encode(body)
	}
}

// EncodeProvidersResponse returns an encoder for responses returned by the
// inventory providers endpoint.
func EncodeProvidersResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		res, _ := v.([]*inventory.XsfcProvider)
		enc := encoder(ctx, w)
		body := NewProvidersResponseBody(res)
		w.WriteHeader(http.StatusOK)
		return enc.Encode(body)
	}
}

// EncodeConsumersResponse returns an encoder for responses returned by the
// inventory consumers endpoint.
func EncodeConsumersResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		res, _ := v.([]*inventory.XsfcConsumer)
		enc := encoder(ctx, w)
		body := NewConsumersResponseBody(res)
		w.WriteHeader(http.StatusOK)
		return enc.Encode(body)
	}
}

// EncodeInjectionsResponse returns an encoder for responses returned by the
// inventory injections endpoint.
func EncodeInjectionsResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		res, _ := v.([]*inventory.XsfcInjection)
		enc := encoder(ctx, w)
		body := NewInjectionsResponseBody(res)
		w.WriteHeader(http.StatusOK)
		return enc.Encode(body)
	}
}

// EncodeAccountsResponse returns an encoder for responses returned by the
// inventory accounts endpoint.
func EncodeAccountsResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		res, _ := v.([]*inventory.XsfcAccount)
		enc := encoder(ctx, w)
		body := NewAccountsResponseBody(res)
		w.WriteHeader(http.StatusOK)
		return enc.Encode(body)
	}
}

// EncodeAccountsByConsumerResponse returns an encoder for responses returned
// by the inventory accountsByConsumer endpoint.
func EncodeAccountsByConsumerResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		res, _ := v.([]*inventory.XsfcAccount)
		enc := encoder(ctx, w)
		body := NewAccountsByConsumerResponseBody(res)
		w.WriteHeader(http.StatusOK)
		return enc.Encode(body)
	}
}

// DecodeAccountsByConsumerRequest returns a decoder for requests sent to the
// inventory accountsByConsumer endpoint.
func DecodeAccountsByConsumerRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (*inventory.AccountsByConsumerPayload, error) {
	return func(r *http.Request) (*inventory.AccountsByConsumerPayload, error) {
		var payload *inventory.AccountsByConsumerPayload
		var (
			namespace string
			name      string

			params = mux.Vars(r)
		)
		namespace = params["namespace"]
		name = params["name"]
		payload = NewAccountsByConsumerPayload(namespace, name)

		return payload, nil
	}
}

// EncodeManifestsRequestingInjectionResponse returns an encoder for responses
// returned by the inventory manifestsRequestingInjection endpoint.
func EncodeManifestsRequestingInjectionResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		res, _ := v.([]*inventory.XsfcManifest)
		enc := encoder(ctx, w)
		body := NewManifestsRequestingInjectionResponseBody(res)
		w.WriteHeader(http.StatusOK)
		return enc.Encode(body)
	}
}

// EncodeHealthzResponse returns an encoder for responses returned by the
// inventory healthz endpoint.
func EncodeHealthzResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		res, _ := v.(string)
		enc := encoder(ctx, w)
		body := res
		w.WriteHeader(http.StatusOK)
		return enc.Encode(body)
	}
}

// EncodeReadyzResponse returns an encoder for responses returned by the
// inventory readyz endpoint.
func EncodeReadyzResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		res, _ := v.(string)
		enc := encoder(ctx, w)
		body := res
		w.WriteHeader(http.StatusOK)
		return enc.Encode(body)
	}
}

// marshalInventoryXsfcModuleToXsfcModuleResponse builds a value of type
// *XsfcModuleResponse from a value of type *inventory.XsfcModule.
func marshalInventoryXsfcModuleToXsfcModuleResponse(v *inventory.XsfcModule) *XsfcModuleResponse {
	res := &XsfcModuleResponse{
		Name:    v.Name,
		Version: v.Version,
	}
	if v.Types != nil {
		res.Types = make([]string, len(v.Types))
		for i, val := range v.Types {
			res.Types[i] = val
		}
	}
	if v.Capabilities != nil {
		res.Capabilities = make([]string, len(v.Capabilities))
		for i, val := range v.Capabilities {
			res.Capabilities[i] = val
		}
	}

	return res
}

// marshalInventoryXsfcProviderToXsfcProviderResponse builds a value of type
// *XsfcProviderResponse from a value of type *inventory.XsfcProvider.
func marshalInventoryXsfcProviderToXsfcProviderResponse(v *inventory.XsfcProvider) *XsfcProviderResponse {
	res := &XsfcProviderResponse{
		Type:      v.Type,
		Name:      v.Name,
		Namespace: v.Namespace,
		Kind:      v.Kind,
		Resource:  v.Resource,
		Module:    v.Module,
	}

	return res
}

// marshalInventoryXsfcConsumerToXsfcConsumerResponse builds a value of type
// *XsfcConsumerResponse from a value of type *inventory.XsfcConsumer.
func marshalInventoryXsfcConsumerToXsfcConsumerResponse(v *inventory.XsfcConsumer) *XsfcConsumerResponse {
	res := &XsfcConsumerResponse{
		Type:      v.Type,
		Name:      v.Name,
		Namespace: v.Namespace,
		Kind:      v.Kind,
		Resource:  v.Resource,
	}
	if v.RequestedTypes != nil {
		res.RequestedTypes = make([]string, len(v.RequestedTypes))
		for i, val := range v.RequestedTypes {
			res.RequestedTypes[i] = val
		}
	}

	return res
}

// marshalInventoryXsfcInjectionToXsfcInjectionResponse builds a value of type
// *XsfcInjectionResponse from a value of type *inventory.XsfcInjection.
func marshalInventoryXsfcInjectionToXsfcInjectionResponse(v *inventory.XsfcInjection) *XsfcInjectionResponse {
	res := &XsfcInjectionResponse{
		ConsumerNamespace: v.ConsumerNamespace,
		ConsumerName:      v.ConsumerName,
		ConsumerKind:      v.ConsumerKind,
		Container:         v.Container,
		Mode:              v.Mode,
		Status:            v.Status,
		SourceManifest:    v.SourceManifest,
	}
	if v.RequestedTypes != nil {
		res.RequestedTypes = make([]string, len(v.RequestedTypes))
		for i, val := range v.RequestedTypes {
			res.RequestedTypes[i] = val
		}
	}

	return res
}

// marshalInventoryXsfcAccountToXsfcAccountResponse builds a value of type
// *XsfcAccountResponse from a value of type *inventory.XsfcAccount.
func marshalInventoryXsfcAccountToXsfcAccountResponse(v *inventory.XsfcAccount) *XsfcAccountResponse {
	res := &XsfcAccountResponse{
		Name:              v.Name,
		Namespace:         v.Namespace,
		Type:              v.Type,
		ConsumerNamespace: v.ConsumerNamespace,
		ConsumerName:      v.ConsumerName,
		ProviderName:      v.ProviderName,
		ProviderNamespace: v.ProviderNamespace,
		CreatedBy:         v.CreatedBy,
	}

	return res
}

// marshalInventoryXsfcManifestToXsfcManifestResponse builds a value of type
// *XsfcManifestResponse from a value of type *inventory.XsfcManifest.
func marshalInventoryXsfcManifestToXsfcManifestResponse(v *inventory.XsfcManifest) *XsfcManifestResponse {
	res := &XsfcManifestResponse{
		APIVersion: v.APIVersion,
		Kind:       v.Kind,
		Name:       v.Name,
		Namespace:  v.Namespace,
	}
	if v.RequestedTypes != nil {
		res.RequestedTypes = make([]string, len(v.RequestedTypes))
		for i, val := range v.RequestedTypes {
			res.RequestedTypes[i] = val
		}
	}
	if v.Annotations != nil {
		res.Annotations = make(map[string]string, len(v.Annotations))
		for key, val := range v.Annotations {
			tk := key
			tv := val
			res.Annotations[tk] = tv
		}
	}
	if v.Labels != nil {
		res.Labels = make(map[string]string, len(v.Labels))
		for key, val := range v.Labels {
			tk := key
			tv := val
			res.Labels[tk] = tv
		}
	}

	return res
}