# cmd/apigen/gen/http/openapi3.yaml
openapi: 3.0.3
info:
    title: XSFC Resource Operator API
    description: REST API for listing providers, consumers, injections, accounts, modules and monitored types.
    version: 0.1.0
servers:
    - url: http://localhost:8088
paths:
...
    /accounts/by-consumer/{namespace}/{name}:
        get:
            tags:
                - inventory
            summary: accountsByConsumer inventory
            operationId: inventory#accountsByConsumer
            parameters:
                - name: namespace
                  in: path
                  required: true
                  schema:
                    type: string
                    example: Sed consequatur est.
                  example: Praesentium expedita magnam aut voluptates fugiat.
                - name: name
                  in: path
                  required: true
                  schema:
                    type: string
                    example: Quo nihil.
                  example: Maxime rerum nulla suscipit adipisci.
            responses:
                "200":
                    description: OK response.
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/XsfcAccount'
                                example:
                                    - consumerName: Commodi pariatur et sint accusantium suscipit eum.
                                      consumerNamespace: Consequatur omnis impedit dolor nam vel voluptates.
                                      createdBy: Eum voluptas in inventore.
                                      name: Accusantium esse omnis alias.
                                      namespace: Omnis laudantium eaque nobis molestiae animi.
                                      providerName: Quae quas et rerum laboriosam.
                                      providerNamespace: Sit est et saepe corporis id delectus.
                                      type: Nam similique delectus quae eius.
                                    - consumerName: Commodi pariatur et sint accusantium suscipit eum.
                                      consumerNamespace: Consequatur omnis impedit dolor nam vel voluptates.
                                      createdBy: Eum voluptas in inventore.
                                      name: Accusantium esse omnis alias.
                                      namespace: Omnis laudantium eaque nobis molestiae animi.
                                      providerName: Quae quas et rerum laboriosam.
                                      providerNamespace: Sit est et saepe corporis id delectus.
                                      type: Nam similique delectus quae eius.
                                    - consumerName: Commodi pariatur et sint accusantium suscipit eum.
                                      consumerNamespace: Consequatur omnis impedit dolor nam vel voluptates.
                                      createdBy: Eum voluptas in inventore.
                                      name: Accusantium esse omnis alias.
                                      namespace: Omnis laudantium eaque nobis molestiae animi.
                                      providerName: Quae quas et rerum laboriosam.
                                      providerNamespace: Sit est et saepe corporis id delectus.
                                      type: Nam similique delectus quae eius.
                                    - consumerName: Commodi pariatur et sint accusantium suscipit eum.
                                      consumerNamespace: Consequatur omnis impedit dolor nam vel voluptates.
                                      createdBy: Eum voluptas in inventore.
                                      name: Accusantium esse omnis alias.
                                      namespace: Omnis laudantium eaque nobis molestiae animi.
                                      providerName: Quae quas et rerum laboriosam.
                                      providerNamespace: Sit est et saepe corporis id delectus.
                                      type: Nam similique delectus quae eius.
                            example:
                                - consumerName: Commodi pariatur et sint accusantium suscipit eum.
                                  consumerNamespace: Consequatur omnis impedit dolor nam vel voluptates.
                                  createdBy: Eum voluptas in inventore.
                                  name: Accusantium esse omnis alias.
                                  namespace: Omnis laudantium eaque nobis molestiae animi.
                                  providerName: Quae quas et rerum laboriosam.
                                  providerNamespace: Sit est et saepe corporis id delectus.
                                  type: Nam similique delectus quae eius.
                                - consumerName: Commodi pariatur et sint accusantium suscipit eum.
                                  consumerNamespace: Consequatur omnis impedit dolor nam vel voluptates.
                                  createdBy: Eum voluptas in inventore.
                                  name: Accusantium esse omnis alias.
                                  namespace: Omnis laudantium eaque nobis molestiae animi.
                                  providerName: Quae quas et rerum laboriosam.
                                  providerNamespace: Sit est et saepe corporis id delectus.
                                  type: Nam similique delectus quae eius.
...
components:
    schemas:
        XsfcAccount:
            type: object
            properties:
                consumerName:
                    type: string
                    example: Sunt laboriosam eius doloremque enim quam vero.
                consumerNamespace:
                    type: string
                    example: Maxime repudiandae.
                createdBy:
                    type: string
                    example: Ea quisquam quaerat laudantium nemo eligendi harum.
                name:
                    type: string
                    example: Atque iste odit fugiat.
                namespace:
                    type: string
                    example: Quia sunt repellendus ab rerum.
                providerName:
                    type: string
                    example: Eveniet repellendus est totam exercitationem alias eveniet.
                providerNamespace:
                    type: string
                    example: Architecto unde eos.
                type:
                    type: string
                    example: Modi repellat enim.
            example:
                consumerName: Est et numquam qui iste.
                consumerNamespace: Quaerat incidunt dolor autem.
                createdBy: Optio error voluptatem qui.
                name: Itaque rerum sunt vero dolores.
                namespace: Voluptates suscipit.
                providerName: Consectetur officiis eaque explicabo incidunt.
                providerNamespace: Impedit sint.
                type: Totam eaque in voluptas quae in consectetur.
...